refactor: don't inject fragments with dagger, use std Activity to create fragments
+ i like the fact that the controllers were DInjected to the graph
- injecting fragments is nice only in theory:
  * i worry about storing and leaking of those componets (what if the framework decides to kill one of them, is something changed about it's state that would make it non-resuable. The graph wouldn't know of that?)
  * fragments are not like POJOs, they're behavior is not set in stone. i'm not too comfortable DInjecting such objects
  * also was facing issue with getting hold of activity in the action bar onAdd method, cause the graph was injecting the frag even before the state of attaching, so the method didn't have access to activity

it's crazy to cargo cult program and force fragments to also be injected through the OGraph
my objective is to be able to unit test... and fast.
the strategy of using the frags/activities still enable that. T
hey take in dependencies through the constructor which totally works for unit testing
i don't care too much about testing the activity/fragment in the same way as i do the controllers;
because they are NOT the same. i would use espresso which would deploy an apk etc and test that... on a real device
cause fragments/activities are framework reliant... the true way to test them is with the framework i.e. on a real device